Move unique validator in separate library
