+++ common-libraries/pom.xml
@@ -6,6 +6,7 @@
     <modules>
         <module>common-exception-handling</module>
         <module>common-exceptions</module>
+        <module>common-validator</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
