+++ specialization-server/specialization-server-service/pom.xml
@@ -22,11 +22,6 @@
             <artifactId>specialization-server-repository</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>by.kovzov.uis.common</groupId>
-            <artifactId>common-exceptions</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
 
         <dependency>
             <groupId>org.projectlombok</groupId>
