+++ specialization-server/specialization-server-service/src/main/java/by/kovzov/uis/specialization/service/impl/SpecializationServiceImpl.java
@@ -8,16 +8,15 @@ import static by.kovzov.uis.specialization.repository.specification.Specializati
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import by.kovzov.uis.common.exception.NotFoundException;
+import by.kovzov.uis.common.validator.unique.UniqueValidationService;
 import by.kovzov.uis.specialization.dto.SpecializationDto;
 import by.kovzov.uis.specialization.dto.SpecializationRequestDto;
 import by.kovzov.uis.specialization.repository.api.SpecializationRepository;
 import by.kovzov.uis.specialization.repository.entity.Specialization;
 import by.kovzov.uis.specialization.service.api.SpecializationService;
-import by.kovzov.uis.specialization.service.api.UniqueValidationService;
 import by.kovzov.uis.specialization.service.mapper.SpecializationMapper;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
